+++ lib/animations/scaled_slide_animation.dart
@@ -1,47 +1,50 @@
 import 'package:flutter/material.dart';
 
-/// [FadedSlideAnimation] is build using [FadeTransition] and [SlideTransition]
+/// [ScaledSlideAnimation] is build using [ScaleTransition] and [SlideTransition]
 
 /// fade and slide animation wrapper widget
 /// just wrap the child with this widget and widget will be animated
 /// [child] : the widget to be animated
 /// [beginOffset] : the starting position of widget to be animated
 /// [endOffset] : the ending position of widget to be animated
-/// [fadeDuration] : duration of animation of fade (default: Duration(milliseconds: 1000))
+/// [scaleDuration] : duration of animation of scale (default: Duration(milliseconds: 500))
 /// [slideDuration] : duration of animation of slide (default: Duration(milliseconds: 500))
-/// [fadeCurve] : curve of the fade animation (default: [Curves.decelerate])
+/// [scaleCurve] : curve of the scale animation (default: [Curves.decelerate])
 /// [slideCurve] : curve of the slide animation (default: [Curves.easeInOutSine])
+/// [alignment] : alignment of the scale animation (default: [Alignment.center])
 /// [textDirection] : direction of the text for slide animation (default: [TextDirection.ltr])
-class FadedSlideAnimation extends StatefulWidget {
+class ScaledSlideAnimation extends StatefulWidget {
   final Widget child;
   final Offset beginOffset;
   final Offset endOffset;
-  final Duration fadeDuration;
+  final Duration scaleDuration;
   final Duration slideDuration;
-  final Curve fadeCurve;
+  final Curve scaleCurve;
   final Curve slideCurve;
+  final Alignment alignment;
   final TextDirection textDirection;
 
-  FadedSlideAnimation({
+  ScaledSlideAnimation({
     Key? key,
     required this.child,
     required this.beginOffset,
     required this.endOffset,
-    this.fadeDuration = const Duration(milliseconds: 1000),
+    this.scaleDuration = const Duration(milliseconds: 500),
     this.slideDuration = const Duration(milliseconds: 500),
-    this.fadeCurve = Curves.decelerate,
+    this.scaleCurve = Curves.decelerate,
     this.slideCurve = Curves.easeInOutSine,
+    this.alignment = Alignment.center,
     this.textDirection = TextDirection.ltr,
   }) : super(key: key);
 
   @override
-  _FadedSlideAnimationState createState() => _FadedSlideAnimationState();
+  _ScaledSlideAnimationState createState() => _ScaledSlideAnimationState();
 }
 
-/// [TickerProviderStateMixin] is used because two animation controllers are being used for making [FadedSlideAnimation].
-class _FadedSlideAnimationState extends State<FadedSlideAnimation>
+/// [TickerProviderStateMixin] is used because two animation controllers are being used for making [ScaledSlideAnimation].
+class _ScaledSlideAnimationState extends State<ScaledSlideAnimation>
     with TickerProviderStateMixin {
-  late AnimationController _fadeController;
+  late AnimationController _scaleController;
   late CurvedAnimation _animation;
   late AnimationController _slideController;
   late Animation<Offset> _offsetAnimation;
@@ -50,15 +53,15 @@ class _FadedSlideAnimationState extends State<FadedSlideAnimation>
   @override
   void initState() {
     super.initState();
-    _fadeController = AnimationController(
-      duration: widget.fadeDuration,
+    _scaleController = AnimationController(
+      duration: widget.scaleDuration,
       vsync: this,
     )..addListener(() {
         setState(() {});
       });
     _animation =
-        CurvedAnimation(parent: _fadeController, curve: widget.fadeCurve);
-    _fadeController.forward();
+        CurvedAnimation(parent: _scaleController, curve: widget.scaleCurve);
+    _scaleController.forward();
     _slideController = AnimationController(
       duration: widget.slideDuration,
       vsync: this,
@@ -73,15 +76,16 @@ class _FadedSlideAnimationState extends State<FadedSlideAnimation>
   /// both the animation controllers are being disposed for safety
   @override
   void dispose() {
-    _fadeController.dispose();
+    _scaleController.dispose();
     _slideController.dispose();
     super.dispose();
   }
 
   @override
   Widget build(BuildContext context) {
-    return FadeTransition(
-      opacity: _animation,
+    return ScaleTransition(
+      scale: _animation,
+      alignment: widget.alignment,
       child: SlideTransition(
         position: _offsetAnimation,
         child: widget.child,